+++ src/effitor/plugins/comp/code/EtCodeElement.ts
@@ -13,7 +13,4 @@ export class EtCodeElement extends EtComponentElement {
     static elName = C.TAG
     static cssText = cssText;
 
-    connectedCallback(this: EtCodeElement): void {
-        this.className = C.TAG
-    }
 }