+++ vite.config.ts
@@ -1,10 +1,34 @@
 import { fileURLToPath, URL } from "url";
 import { defineConfig } from "vite";
 
-export default defineConfig({
-    resolve: {
-        alias: {
-            '@': fileURLToPath(new URL('./src', import.meta.url))
+export default defineConfig(({ command, mode }) => {
+    if (command === 'serve') {
+        // dev
+        return {
+            resolve: {
+                alias: {
+                    '@': fileURLToPath(new URL('./src', import.meta.url))
+                },
+            },
+
+        }
+    }
+    else {
+        // build
+        return {
+            resolve: {
+                alias: {
+                    '@': fileURLToPath(new URL('./temp/src', import.meta.url))
+                },
+            },
+            build: {
+                minify: false,
+                lib: {
+                    entry: 'temp/src/effitor/index.js',
+                    fileName: 'index',
+                    formats: ['es'],
+                },
+            }
         }
     }
 })